Fix build timeout due to dataset download issues
