+++ pyro/contrib/examples/util.py
@@ -2,9 +2,23 @@ from __future__ import absolute_import, division, print_function
 
 import sys
 
+import torchvision
 import torchvision.datasets as datasets
 from torch.utils.data import DataLoader
 from torchvision import transforms
+from torchvision.datasets import MNIST
+
+from pyro.util import patch_dependency
+
+
+@patch_dependency('torchvision.datasets.MNIST', torchvision)
+class _MNIST(MNIST):
+    urls = [
+        "https://d2fefpcigoriu7.cloudfront.net/datasets/mnist/train-images-idx3-ubyte.gz",
+        "https://d2fefpcigoriu7.cloudfront.net/datasets/mnist/train-labels-idx1-ubyte.gz",
+        "https://d2fefpcigoriu7.cloudfront.net/datasets/mnist/t10k-images-idx3-ubyte.gz",
+        "https://d2fefpcigoriu7.cloudfront.net/datasets/mnist/t10k-labels-idx1-ubyte.gz",
+    ]
 
 
 def get_data_loader(dataset_name,
