+++ pyro/distributions/torch_patch.py
@@ -2,27 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 import torch
 
-
-def _patch(target):
-    parts = target.split('.')
-    assert parts[0] == 'torch'
-    module = torch
-    for part in parts[1:-1]:
-        module = getattr(module, part)
-    name = parts[-1]
-    old_fn = getattr(module, name)
-    old_fn = getattr(old_fn, '_pyro_unpatched', old_fn)  # ensure patching is idempotent
-
-    def decorator(new_fn):
-        new_fn.__name__ = name
-        new_fn._pyro_unpatched = old_fn
-        setattr(module, name, new_fn)
-        return new_fn
-
-    return decorator
+from pyro.util import patch_dependency
 
 
-@_patch('torch._standard_gamma')
+@patch_dependency('torch._standard_gamma')
 def _torch_standard_gamma(concentration):
     unpatched_fn = _torch_standard_gamma._pyro_unpatched
     if concentration.is_cuda:
@@ -30,14 +13,14 @@ def _torch_standard_gamma(concentration):
     return unpatched_fn(concentration)
 
 
-@_patch('torch.distributions.gamma._standard_gamma')
+@patch_dependency('torch.distributions.gamma._standard_gamma')
 def _standard_gamma(concentration):
     if concentration.is_cuda:
         return concentration.cpu()._standard_gamma().cuda(concentration.get_device())
     return concentration._standard_gamma()
 
 
-@_patch('torch._dirichlet_grad')
+@patch_dependency('torch._dirichlet_grad')
 def _torch_dirichlet_grad(x, concentration, total):
     unpatched_fn = _torch_dirichlet_grad._pyro_unpatched
     if x.is_cuda:
@@ -46,7 +29,7 @@ def _torch_dirichlet_grad(x, concentration, total):
 
 
 # This can be removed when super(...).__init__() is added upstream
-@_patch('torch.distributions.transforms.Transform.__init__')
+@patch_dependency('torch.distributions.transforms.Transform.__init__')
 def _Transform__init__(self, cache_size=0):
     self._cache_size = cache_size
     self._inv = None
@@ -59,7 +42,7 @@ def _Transform__init__(self, cache_size=0):
     super(torch.distributions.transforms.Transform, self).__init__()
 
 
-@_patch('torch.linspace')
+@patch_dependency('torch.linspace')
 def _torch_linspace(*args, **kwargs):
     unpatched_fn = _torch_linspace._pyro_unpatched
     template = torch.Tensor()
@@ -72,7 +55,7 @@ def _torch_linspace(*args, **kwargs):
     return ret
 
 
-@_patch('torch.einsum')
+@patch_dependency('torch.einsum')
 def _einsum(equation, operands):
     # work around torch.einsum performance issues
     # see https://github.com/pytorch/pytorch/issues/10661
